+++ pkg/domen/models/entities/category_model.go
@@ -1,12 +1,12 @@
 package entities
 
 type Category struct {
-	ID             int32           `json:"id"`
+	ID             int64           `json:"id"`
 	Name           string          `json:"name"`
 	ParentCategory *ParentCategory `json:"parent_category"`
 }
 
 type ParentCategory struct {
-	ID   int32  `json:"id"`
+	ID   int64  `json:"id"`
 	Name string `json:"name"`
 }
